requirements.txt
fastapi
uvicorn[standard]
aiogram==3.13.0
python-dotenv